# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.structureAreaRequest:
    properties:
      area_level_id:
        type: integer
      code:
        type: string
      name:
        type: string
      status:
        type: string
      store_id:
        type: integer
    required:
    - area_level_id
    - code
    - name
    - store_id
    type: object
  handlers.structureBuildingRequest:
    properties:
      code:
        type: string
      name:
        type: string
      status:
        type: string
      store_id:
        type: integer
    required:
    - code
    - name
    - store_id
    type: object
  handlers.structureFloorRequest:
    properties:
      building_id:
        type: integer
      code:
        type: string
      floor_plan_image_url:
        type: string
      name:
        type: string
      status:
        type: string
    required:
    - building_id
    - code
    - name
    type: object
...
  handlers.swaggerEnvelope:
    type: object
  handlers.swaggerMessageResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  structure.Area:
    properties:
      area_level_id:
        type: integer
      code:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: string
      store_id:
        type: integer
      updated_at:
        type: string
    type: object
  structure.Building:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: string
      store_id:
        type: integer
      updated_at:
        type: string
    type: object
  structure.Floor:
    properties:
      building_id:
        type: integer
      code:
        type: string
      created_at:
        type: string
      floor_plan_image_url:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:5180
info:
  contact: {}
  description: Backend API for the MI property management system
  title: MI Management Information System API
  version: "1.0"
paths:
...
  /structure/areas:
    get:
      consumes:
      - application/json
      description: Returns areas filtered by store.
      parameters:
      - description: Store ID
        in: query
        name: store_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                areas:
                  items:
                    $ref: '#/definitions/structure.Area'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: List areas
      tags:
      - Structure
    post:
      consumes:
      - application/json
      description: Creates a new area record.
      parameters:
      - description: Area request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.structureAreaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                area:
                  $ref: '#/definitions/structure.Area'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: Create area
      tags:
      - Structure
  /structure/areas/{id}:
    put:
      consumes:
      - application/json
      description: Updates an existing area record.
      parameters:
      - description: Area ID
        in: path
        name: id
        required: true
        type: integer
      - description: Area request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.structureAreaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                area:
                  $ref: '#/definitions/structure.Area'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: Update area
      tags:
      - Structure
  /structure/buildings:
    get:
      consumes:
      - application/json
      description: Returns buildings filtered by store.
      parameters:
      - description: Store ID
        in: query
        name: store_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                buildings:
                  items:
                    $ref: '#/definitions/structure.Building'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: List buildings
      tags:
      - Structure
    post:
      consumes:
      - application/json
      description: Creates a new building record.
      parameters:
      - description: Building request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.structureBuildingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                building:
                  $ref: '#/definitions/structure.Building'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: Create building
      tags:
      - Structure
  /structure/buildings/{id}:
    put:
      consumes:
      - application/json
      description: Updates an existing building record.
      parameters:
      - description: Building ID
        in: path
        name: id
        required: true
        type: integer
      - description: Building request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.structureBuildingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                building:
                  $ref: '#/definitions/structure.Building'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: Update building
      tags:
      - Structure
...
  /structure/floors/{id}:
    put:
      consumes:
      - application/json
      description: Updates an existing floor record.
      parameters:
      - description: Floor ID
        in: path
        name: id
        required: true
        type: integer
      - description: Floor request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.structureFloorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                floor:
                  $ref: '#/definitions/structure.Floor'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: Update floor
      tags:
      - Structure
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"